fix: finish daily start feature

+++ backend/src/main/java/com/stacknstock/backend/domain/game/dto/PortfolioResponse.java
@@ -1,7 +1,9 @@
 package com.stacknstock.backend.domain.game.dto;
 
+import com.stacknstock.backend.domain.trade.dto.TradeLogResponse;
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Schema(description = "현재 포트폴리오 조회 응답")
@@ -14,10 +16,10 @@ public record PortfolioResponse(
         Integer currentDayNo,
 
         @Schema(description = "보유 현금", example = "9800000")
-        Long cashBalance,
+        BigDecimal cashBalance,
 
         @Schema(description = "총 평가 자산", example = "10250000")
-        Long totalAssetValue,
+        BigDecimal totalAssetValue,
 
         @Schema(description = "보유 종목 리스트")
         List<PortfolioStockResponse> holdings,
